cygwin build fixes
and shrink some too-long lines

+++ src/flash/nand/at91sam9.c
@@ -414,12 +414,16 @@ static int at91sam9_read_page(struct nand_device *nand, uint32_t page,
 				// attempt recovery
 				uint32_t parity;
 
-				target_read_u32(target, info->ecc + AT91C_ECCx_PR, &parity);
+				target_read_u32(target,
+						info->ecc + AT91C_ECCx_PR,
+						&parity);
 				uint32_t word = (parity & 0x0000FFF0) >> 4;
 				uint32_t bit = parity & 0x0F;
 
 				data[word] ^= (0x1) << bit;
-				LOG_INFO("Data word %d, bit %d corrected.", word, bit);
+				LOG_INFO("Data word %d, bit %d corrected.",
+						(unsigned) word,
+						(unsigned) bit);
 			}
 		}
